test (integrator): path tracer simple test

+++ src/Integrator/MonteCarloIntegrator.cpp
@@ -11,7 +11,7 @@
 
 #include "MonteCarloIntegrator.h"
 #include "../Camera/TestCamera.h"
- 
+
 MonteCarloIntegrator::MonteCarloIntegrator(std::shared_ptr<Camera> camera, std::unique_ptr<Film> film, std::unique_ptr<TileGenerator> tileGenerator, std::shared_ptr<Sampler> sampler, int spp) : Integrator(camera, std::move(film), std::move(tileGenerator)), sampler(sampler), spp(spp)
 {
 }
@@ -29,7 +29,7 @@ void MonteCarloIntegrator::render(std::shared_ptr<Scene> scene)
             {
                 Point2d NDC(2.0 * (x + sampler->sample()) / filmWidth - 1, -2.0 * (y + sampler->sample()) / filmHeight + 1);
                 auto L = Li(cam.generateRay(NDC), scene);
-                film->deposit(Point2i(x, y), L);
+                film->deposit(Point2d(x, y), L);
             }
         }
     }
@@ -39,4 +39,4 @@ double MonteCarloIntegrator::randFloat()
 {
     // Get a random number WITHOUT using MonteCarloIntegrator::sampler
     return rand() * 1.0 / RAND_MAX; // todo: better solution
-} 
+}
